requirements.txt
PyYAML>=6.0
requests>=2.28.0
Pillow>=9.0.0
psutil>=5.9.0
py7zr>=0.20.0
